move config folder to root
