+++ src/config/app.ts
@@ -1,4 +1,4 @@
-import { AppConfig } from "../../types"
+import { AppConfig } from "../types"
 
 const appConfig: AppConfig = {
   title: "hypBox Lab",
